// frontend/alanda-libraries/projects/alanda-common/src/lib/api/models/alandaProject.ts
import { AlandaProcess } from './alandaProcess';
import { ProjectState } from '../../enums/projectState.enum';
import { AlandaMilestone } from './alandaMilestone';
import { AlandaSimplePhase } from './alandaSimplePhase';
import { AlandaHistoryLog } from './alandaHistoryLog';
import { AlandaProperty } from './alandaProperty';
import { AlandaProjectType } from './alandaProjectType';

export interface AlandaProject {
    guid?: number;
    version?: number;
    projectId?: string;
    tag?: string[];
    title?: string;
    status?: ProjectState;
    priority?: number;
    refObjectIdName?: string;
    refObjectType?: string;
    projectType?: string;
    projectTypeIdName?: string;
    subType?: string;
    dueDate?: string;
    pmcProjectType?: AlandaProjectType;
    subtype?: string;
    parents?: AlandaProject[];
    children?: AlandaProject[];
    processes?: AlandaProcess[];
    customerProjectId?: number;
    ownerId?: number;
    details?: string;
    guStatus?: string;
    comment?: string;
    risk?: number;
    refObjectId?: number;
    milestones?: AlandaMilestone[];
    phases?: AlandaSimplePhase[];
    history?: AlandaHistoryLog[];
    displayMetaInfo?: string;
    properties?: AlandaProperty[];
    propertiesMap?: Map<string, any>;
    milestonesMap?: Map<string, any>;
    createDate?: string;
    createUser?: number;
    ownerName?: string;
    updateDate?: string;
    updateUser?: number;
    parentIds?: number[];
    childrenIds?: number[];
    additionalInfo?: Map<string, any>;
    resultStatus?: string;
    resultComment?: string;
    authBase?: string;
    running?: boolean;
    humanReadableId?: string;
}